* dwarf.c (display_debug_loc): Cast section_end - start to long 	for output format.

+++ binutils/dwarf.c
@@ -2627,7 +2627,7 @@ display_debug_loc (struct dwarf_section *section, void *file)
 
   if (start < section_end)
     warn (_("There are %ld unused bytes at the end of section %s\n"),
-	  section_end - start, section->name);
+	  (long) (section_end - start), section->name);
   return 1;
 }
